+++ packages/jest-core/src/runJest.ts
@@ -97,9 +97,8 @@ const processResults = async (
   }
 
   if (testResultsProcessor) {
-    const processor = await requireOrImportModule<TestResultsProcessor>(
-      testResultsProcessor,
-    );
+    const processor =
+      await requireOrImportModule<TestResultsProcessor>(testResultsProcessor);
     runResults = await processor(runResults);
   }
   if (isJSON) {